Remove extra space from command name
When `$CMD` was switched to a bash list instead of a string, bash no longer stripped off the extra '` `' character.  Thus, Singularity was attempting to execute a binary named ` foo` instead of `foo`.

I was able to recreate this by logging in to a worker node with a running glidein, making a copy of the job wrapper script, and applying the patch by hand.  I was able to:
* Reproduce the errors reported by Diego in https://ggus.eu/index.php?mode=ticket_info&ticket_id=128469,
* Verify that this fix took care of the reported error.
* Verify that arguments with spaces in them appear to work.

Mystery solved, hopefully!